+++ CHANGELOG.md
@@ -6,6 +6,13 @@ This is a summary of new features and bugfixes. Read the README to learn how to 
 ## Known issues
 * SDL always merges joycons behind the scene into a single controller. JSM is not able to split them. Use legacy-JSL version to use this feature.
 
+## 3.5.4
+
+Adding a highly requested new feature!
+
+### Features
+* Added Diagonal press as an alternate way to have a binding to 2 buttons. See README section
+
 ## 3.5.3
 
 ### Features